test(source/firestore): consolidate tests into firestore_test.go
